+++ conjuror/plans/plan_generator.py
@@ -17,7 +17,7 @@ from pydicom.uid import generate_uid
 from conjuror.images.layers import ArrayLayer
 from conjuror.images.simulators import Imager, Simulator
 from conjuror.plans.beam import Beam
-from conjuror.plans.machine import TMachine, MachineSpecs, MachineBase
+from conjuror.plans.machine import TMachine, MachineSpecs
 from conjuror.plans.visualization import plot_fluences
 
 
@@ -135,6 +135,8 @@ class PlanGenerator(Generic[TMachine]):
         self._base_plan = base_plan
         self.ds = plan
         self.machine_name = base_plan.BeamSequence[0].TreatmentMachineName
+        self.beam_manufacturer = base_plan.BeamSequence[0].Manufacturer
+        self.beam_manufacturer_model = base_plan.BeamSequence[0].ManufacturerModelName
 
         # Clear beam sequence, this will be filled with the custom beams
         plan.BeamSequence = DicomSequence()
@@ -153,53 +155,6 @@ class PlanGenerator(Generic[TMachine]):
         ds = pydicom.dcmread(rt_plan_file)
         return cls(ds, **kwargs)
 
-    @classmethod
-    def from_machine(
-        cls,
-        machine: MachineBase,
-        machine_name: str = "Conjuror",
-        plan_label: str = "Conjuror",
-        plan_name: str = "Conjuror",
-        patient_name: str = "Conjuror",
-        patient_id: str = "Conjuror",
-    ) -> Self:
-        """Create a plan for a target machine type.
-
-        Parameters
-        ----------
-        machine : MachineBase
-            The target machine.
-        machine_name : str
-            The target machine name.
-        plan_label : str
-            The label of the new plan.
-        plan_name : str
-            The name of the new plan.
-        patient_name : str, optional
-            The name of the patient. If not provided, it will be taken from the RTPLAN file.
-        patient_id : str, optional
-            The ID of the patient. If not provided, it will be taken from the RTPLAN file."""
-        base_plan = Dataset()
-
-        # Transfer syntax UID (Implicit VR Little Endian: Default Transfer Syntax for DICOM)
-        # https://dicom.nema.org/medical/dicom/current/output/chtml/part06/chapter_a.html
-        base_plan.is_implicit_VR = True
-        base_plan.is_little_endian = True
-
-        # General tags required on the base plan
-        base_plan.Modality = "RTPLAN"
-        base_plan.PatientName = patient_name
-        base_plan.PatientID = patient_id
-        base_plan.Manufacturer = "Conjuror"
-
-        # Machine type specific tags required on the base plan
-        sop, beam, tolerance_table = _get_datasets_from_machine_type(machine)
-        beam.TreatmentMachineName = machine_name
-        base_plan.SOPClassUID = sop
-        base_plan.BeamSequence = (beam,)
-        base_plan.ToleranceTableSequence = (tolerance_table,)
-        return cls(base_plan, plan_label, plan_name)
-
     def add_beam(self, beam: Beam):
         """Add a beam to the plan using the Beam object. Although public,
         this is a low-level method that is used by the higher-level methods like add_open_field_beam.
@@ -210,6 +165,8 @@ class PlanGenerator(Generic[TMachine]):
         # Update the beam
         beam_dataset.BeamNumber = len(self.ds.BeamSequence) + 1
         beam_dataset.TreatmentMachineName = self.machine_name
+        beam_dataset.Manufacturer = self.beam_manufacturer
+        beam_dataset.ManufacturerModelName = self.beam_manufacturer_model
         patient_setup_nr = self.ds.PatientSetupSequence[0].PatientSetupNumber
         beam_dataset.ReferencedPatientSetupNumber = patient_setup_nr
         tolerance_table_nr = self.ds.ToleranceTableSequence[0].ToleranceTableNumber
@@ -322,42 +279,3 @@ def _get_machine_type_from_mlc(mlc: Dataset, machine_specs: MachineSpecs) -> TMa
         raise ValueError("MLC type not supported")
 
     return machine
-
-
-def _get_datasets_from_machine_type(
-    machine: MachineBase,
-) -> tuple[str, Dataset, Dataset]:
-    """This function acts as factory to build the required data set from machine type."""
-    # Local imports are used to avoid circular dependencies.
-    # When a new machine type is added, this factory must be updated accordingly.
-    # This is an intentional design choice: although a plugin/registry pattern could
-    # automate new additions, new machine types are expected to be added rarely,
-    # and maintaining explicit control in this method is preferred.
-
-    # Note: This function does not create realistic datasets. Instead, the datasets
-    # contain only the minimum elements required to be processed by
-    # _get_machine_type_from_mlc. For example, LeafPositionBoundaries for TrueBeam
-    # includes only what is needed to distinguish between HDMLC and Millennium,
-    # whereas for Halcyon it does not exist at all.
-
-    tolerance_table = Dataset()
-    tolerance_table.ToleranceTableNumber = 1
-
-    from conjuror.plans.truebeam import TrueBeamMachine
-    from conjuror.plans.halcyon import HalcyonMachine
-
-    bld = Dataset()
-    if isinstance(machine, TrueBeamMachine):
-        sop = "1.2.840.10008.5.1.4.1.1.481.5"
-        bld.RTBeamLimitingDeviceType = "MLCX"
-        bld.LeafPositionBoundaries = 2 * [-110 if machine.mlc_is_hd else -200]
-    elif isinstance(machine, HalcyonMachine):
-        sop = "1.2.246.352.70.1.70"
-        bld.RTBeamLimitingDeviceType = "MLCX1"
-    else:
-        raise ValueError("Unknown machine type")
-
-    beam = Dataset()
-    beam.BeamLimitingDeviceSequence = (bld,)
-
-    return sop, beam, tolerance_table